replace float type by double

+++ src/main/java/dev/embeddings4j/Vector.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Getter
 public class Vector {
     private int id;
-    private float[] values;
+    private double[] values;
 }
