+++ src/test/java/dev/embeddings4j/internal/BruteForceSearchVectorIndexTest.java
@@ -17,11 +17,11 @@ class BruteForceSearchVectorIndexTest {
 
     @Test
     public void testFindNearest() {
-        vectorIndex.addItem(0, new float[]{1.0f, 2.0f, 3.0f});
-        vectorIndex.addItem(1, new float[]{4.0f, 5.0f, 6.0f});
-        vectorIndex.addItem(2, new float[]{7.0f, 8.0f, 9.0f});
+        vectorIndex.addItem(0, new double[]{1.0f, 2.0f, 3.0f});
+        vectorIndex.addItem(1, new double[]{4.0f, 5.0f, 6.0f});
+        vectorIndex.addItem(2, new double[]{7.0f, 8.0f, 9.0f});
 
-        List<Integer> nearestIndices = vectorIndex.searchNearest(new float[]{1.5f, 2.5f, 3.5f}, 2);
+        List<Integer> nearestIndices = vectorIndex.searchNearest(new double[]{1.5f, 2.5f, 3.5f}, 2);
 
         assertEquals(2, nearestIndices.size());
